refactor: 스티커 상태 분리
